requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2
apscheduler
httpx
